Add coverage subcommand to calculate and display test coverage
Instead of relying on the config file to tell us what tasks/outputs to test, the WDL files in the directory tell us what needs testing and then the config file is checked for associated tests. Currently, there are some edge cases that need handling; 1. Compound outputs and 2. Workflows from the custom_tests directory are included in the coverage calculation, which is not correct.